refactor : InputValidator 별도 클래스 삭제 InputView안에 검증코드 구현
